requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
ecdsa==0.19.1
psycopg2-binary==2.9.6
pyasn1==0.6.1
PyJWT==2.9.0
python-jose==3.3.0
pytz==2025.2
rsa==4.9
six==1.17.0
sqlparse==0.5.3
